commenting out existing and merging into 1 file
